pyproject.toml
[tool.poetry]
name = "lamoom"
version = "0.1.33"
description = ""
authors = ["Lamoom Engineering Team <engineering@lamoom.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
tiktoken = ">=0.5.2"
pyyaml = "^6.0.1"
openai = "^1.65.1"
google-generativeai = "^0.7.2"
anthropic = "^0.31.2"
httpx = "^0.27.2"

[tool.poetry.dev-dependencies]
poetry = "^1.7.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pre-commit = "^3.6.0"
autopep8 = "^2.0.4"
python-dotenv = "^1.0.1"
twine = "^5.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
